netbeans Docker

+++ build.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-rm screenlog.0
+[ -f screenlog.0 ]&&rm screenlog.0
 screen -L sh -c "tar -C context -cf - .|docker build --progress=plain -t andzuc/alpine-netbeans -"
